+++ runbot_merge/__manifest__.py
@@ -8,6 +8,7 @@
 
         'data/merge_cron.xml',
         'views/res_partner.xml',
+        'views/runbot_merge_project.xml',
         'views/mergebot.xml',
         'views/templates.xml',
     ],
