+++ runbot_merge/models/__init__.py
@@ -1,2 +1,3 @@
 from . import res_partner
+from . import project
 from . import pull_requests
