+++ runbot_merge/views/runbot_merge_project.xml
@@ -1,0 +1,51 @@
+<odoo>
+    <record id="runbot_merge_form_project" model="ir.ui.view">
+        <field name="name">Project Form</field>
+        <field name="model">runbot_merge.project</field>
+        <field name="arch" type="xml">
+            <form>
+                <header>
+                    <button type="object" name="action_freeze" string="Freeze" class="oe_highlight"/>
+                </header>
+                <sheet>
+                    <div class="oe_title">
+                        <h1><field name="name" placeholder="Name"/></h1>
+                    </div>
+                    <group>
+                        <group>
+                            <field name="github_prefix" string="bot name"/>
+                        </group>
+                    </group>
+                    <group>
+                        <group>
+                            <field name="github_token"/>
+                            <field name="secret"/>
+                        </group>
+                        <group>
+                            <field name="ci_timeout"/>
+                            <field name="batch_limit"/>
+                        </group>
+                    </group>
+
+                    <separator string="Repositories"/>
+                    <field name="repo_ids">
+                        <tree>
+                            <field name="sequence" widget="handle"/>
+                            <field name="name"/>
+                            <field name="branch_filter"/>
+                            <field name="status_ids" widget="many2many_tags"/>
+                        </tree>
+                    </field>
+                    <separator string="Branches"/>
+                    <field name="branch_ids">
+                        <tree editable="bottom" decoration-muted="not active">
+                            <field name="sequence" widget="handle" />
+                            <field name="name"/>
+                            <field name="active"/>
+                        </tree>
+                    </field>
+                </sheet>
+            </form>
+        </field>
+    </record>
+</odoo>
